Modify cpp07/ex02 (test negative index) again

+++ cpp07/ex02/main.cpp
@@ -6,7 +6,7 @@
 /*   By: Yoshihiro Kosaka <ykosaka@student.42tok    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/16 15:04:04 by ykosaka           #+#    #+#             */
-/*   Updated: 2023/07/11 23:46:07 by Yoshihiro K      ###   ########.fr       */
+/*   Updated: 2023/07/11 23:49:46 by Yoshihiro K      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -72,9 +72,9 @@ static int	case_numeric(void) {
 	try {
 		// std::clog << "\033[2;3mIn try{} block\033[m" << std::endl;
 		std::cout << "arr_floats[" << static_cast<int>(-arr_floats.size()) 
-			<< "] * arr_nums[" << -arr_nums.size() << "] = " << std::flush;
+			<< "] * arr_nums[" << -static_cast<int>(arr_nums.size()) << "] = " << std::flush;
 		std::cout << multiply(arr_floats[static_cast<int>(-arr_floats.size())], \
-			arr_nums[-arr_nums.size()]) << std::endl;
+			arr_nums[-static_cast<int>(arr_nums.size())]) << std::endl;
 	}
 	catch (std::exception& e) {
 		// std::clog << "\033[2;3mIn catch{} block\033[m" << std::endl;
